[#1257] Add CUBRID support - Fixed some issues

+++ jOOQ-test/src/org/jooq/test/jOOQDB2Test.java
@@ -545,7 +545,7 @@ public class jOOQDB2Test extends jOOQAbstractTest<
     }
 
     @Override
-    protected TableField<XUnusedRecord, Byte> TDirectory_IS_DIRECTORY() {
+    protected TableField<XUnusedRecord, Integer> TDirectory_IS_DIRECTORY() {
         return null;
     }
