+++ jOOQ-test/src/org/jooq/test/jOOQHSQLDBTest.java
@@ -535,7 +535,7 @@ public class jOOQHSQLDBTest extends jOOQAbstractTest<
     }
 
     @Override
-    protected TableField<XUnusedRecord, Byte> TDirectory_IS_DIRECTORY() {
+    protected TableField<XUnusedRecord, Integer> TDirectory_IS_DIRECTORY() {
         return null;
     }
