+++ jOOQ-test/src/org/jooq/test/jOOQOracleTest.java
@@ -611,7 +611,7 @@ public class jOOQOracleTest extends jOOQAbstractTest<
     }
 
     @Override
-    protected TableField<TDirectoryRecord, Byte> TDirectory_IS_DIRECTORY() {
+    protected TableField<TDirectoryRecord, Integer> TDirectory_IS_DIRECTORY() {
         return T_DIRECTORY.IS_DIRECTORY;
     }
