+++ jOOQ-test/src/org/jooq/test/jOOQPostgresTest.java
@@ -544,7 +544,7 @@ public class jOOQPostgresTest extends jOOQAbstractTest<
     }
 
     @Override
-    protected TableField<XUnusedRecord, Byte> TDirectory_IS_DIRECTORY() {
+    protected TableField<XUnusedRecord, Integer> TDirectory_IS_DIRECTORY() {
         return null;
     }
